% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distritos}
\alias{distritos}
\title{Mozambique distritos borders}
\format{
A SpatialPolygonsDataFrame object containing distritos borders of
Mozambique\tabular{ll}{
   \strong{Variables} \tab \strong{Description} \cr
   Shape_Leng \tab Shape length \cr
   Shape_Area \tab Shape area \cr
   ADM2_PT \tab Administrative area level 2 name in Portuguese \cr
   ADM2_PCODE \tab Administrative area level 2 postal code \cr
   ADM2_REF \tab Administrative area level 2 reference \cr
   ADM2ALT1PT \tab Administrative area level 2 alternative Portuguese name 1 \cr
   ADM2ALT2PT \tab Administrative area level 2 alternative Portuguese name 2 \cr
   ADM1_PT \tab Administrative area level 1 name in Portuguese \cr
   ADM1_PCODE \tab Administrative area level 1 postal code \cr
   ADM0_EN \tab Administrative area level 0 name in English \cr
   ADM0_PT \tab Administrative area level 0 name in Portuguese \cr
   ADM0_PCODE \tab Administrative area level 0 postal code \cr
   date \tab Date created \cr
   validOn \tab Data valid on \cr
   validTo \tab Data valid to \cr
}
}
\source{
https://data.humdata.org/dataset/mozambique-administrative-levels-0-3
}
\usage{
distritos
}
\description{
Mozambique distritos borders
}
\examples{
distritos

}
\keyword{datasets}
