% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shapefiles.R
\name{download_shapefiles}
\alias{download_shapefiles}
\title{Download Mozambique shapefiles}
\usage{
download_shapefiles(id = "cod-ab-moz", .unzip = TRUE)
}
\arguments{
\item{id}{Identifier for Mozambique administrative boundaries from HDX.}

\item{.unzip}{Logical. Should zip file be unzipped and extracted? Default to
TRUE}
}
\value{
If \code{.unzip}, list of shapefiles within the downloaded zip file else
path to downloaded zip file. Mozambique subnational administrative
boundaries zip file downloaded to specified folder.
}
\description{
Download Mozambique shapefiles
}
\examples{
download_shapefiles()

}
