% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shapefiles.R
\name{get_districts}
\alias{get_districts}
\title{Retrieve Mozambique districts boundaries}
\usage{
get_districts(layer = get_layer_name(adm = 2))
}
\arguments{
\item{layer}{Name of layer for districts boundaries}
}
\value{
An sf object for the districts borders of Mozambique
}
\description{
Retrieve Mozambique districts boundaries
}
\examples{
get_districts()

}
