% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shapefiles.R
\name{get_layer_name}
\alias{get_layer_name}
\title{Get layer name}
\usage{
get_layer_name(files = download_shapefiles(), adm = NULL)
}
\arguments{
\item{files}{List of shapefiles within a downloaded zip file}

\item{adm}{Integer. Administrative level required. Either \code{0} for country,
\code{1} for provincias, \code{2} for distritos, or \code{3} for postos}
}
\value{
Layer name for specified administrative level boundary
}
\description{
Get layer name
}
\examples{
get_layer_name(adm = 0)

}
