% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shapefiles.R
\name{get_posts}
\alias{get_posts}
\title{Retrieve Mozambique posts boundaries}
\usage{
get_posts(layer = get_layer_name(adm = 3))
}
\arguments{
\item{layer}{Name of layer for posts boundaries}
}
\value{
An sf object for the posts borders of Mozambique
}
\description{
Retrieve Mozambique posts boundaries
}
\examples{
get_posts()

}
