% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shapefiles.R
\name{get_provinces}
\alias{get_provinces}
\title{Retrieve Mozambique provinces boundaries}
\usage{
get_provinces(layer = get_layer_name(adm = 1))
}
\arguments{
\item{layer}{Name of layer for provinces boundaries}
}
\value{
An sf object for the provinces borders of Mozambique
}
\description{
Retrieve Mozambique provinces boundaries
}
\examples{
get_provinces()

}
