% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppl}
\alias{ppl}
\title{Mozambique populated places locations}
\format{
A \code{SpatialPointsDataFrame} object containing populated places
locations of Mozambique

| \strong{Variables} | \strong{Description} |
| \emph{osm_id} | OpenStreetMap identifier |
| \emph{name} | Name of populated place |
| \emph{source} | Source of data for populated place |
| \emph{population} | Population size |
| \emph{is_in} | Populated place is in? |
| \emph{place} | Type of place |
}
\source{
https://data.humdata.org/dataset/hotosm_moz_populated_places
}
\usage{
ppl
}
\description{
Mozambique populated places locations
}
\examples{
ppl

}
\keyword{datasets}
