% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{provincias}
\alias{provincias}
\title{Mozambique provincias borders}
\format{
A SpatialPolygonsDataFrame object containing provincias borders of
Mozambique\tabular{ll}{
   \strong{Variables} \tab \strong{Description} \cr
   Shape_Leng \tab Shape length \cr
   Shape_Area \tab Shape area \cr
   ADM1_PT \tab Administrative area level 1 name in Portuguese \cr
   ADM1_PCODE \tab Administrative area level 1 postal code \cr
   ADM1_REF \tab Administrative area level 1 reference \cr
   ADM1ALT1PT \tab Administrative area level 1 alternative Portuguese name 1 \cr
   ADM1ALT2PT \tab Administrative area level 1 alternative Portuguese name 2 \cr
   ADM0_EN \tab Administrative area level 0 name in English \cr
   ADM0_PT \tab Administrative area level 0 name in Portuguese \cr
   ADM0_PCODE \tab Administrative area level 0 postal code \cr
   date \tab Date created \cr
   validOn \tab Data valid on \cr
   validTo \tab Data valid to \cr
}
}
\source{
https://data.humdata.org/dataset/mozambique-administrative-levels-0-3
}
\usage{
provincias
}
\description{
Mozambique provincias borders
}
\examples{
provincias

}
\keyword{datasets}
