% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{settlements}
\alias{settlements}
\title{Mozambique settlements locations}
\format{
A \code{SpatialPointsDataFrame} object containing settlements locations of
Mozambique

| \strong{Variables} | \strong{Description} |
| \emph{OBJECTID} | Object identifier |
| \emph{FID_1} | Feature identifier |
| \emph{Sett_Name} | Settlement name |
| \emph{F_CLASS} | Feature class |
| \emph{F_DESIG} | Feature designation |
| \emph{ADM1_Old} | Administrative level 1 name - old |
}
\source{
https://data.humdata.org/dataset/mozambique-settlement-shapefiles
}
\usage{
settlements
}
\description{
Mozambique settlements locations
}
\examples{
settlements

}
\keyword{datasets}
