% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{villages}
\alias{villages}
\title{Mozambique villages locations}
\format{
A SpatialPointsDataFrame object containing villages locations of
Mozambique

| \strong{Variables} | \strong{Description} |
| osm_id | OpenStreetMap Identifier |
| name | Name of location |
| source | Source of location data |
| population | Population size |
| is_in | Is in what country location |
| place | Type of location |
}
\source{
https://data.humdata.org/dataset/hotosm_moz_populated_places
}
\usage{
villages
}
\description{
Mozambique villages locations
}
\examples{
villages

}
\keyword{datasets}
