% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{provincias}
\alias{provincias}
\title{Mozambique \emph{provincias} borders}
\format{
A \code{sf} object containing \emph{provincias} borders of Mozambique\tabular{ll}{
   \strong{Variables} \tab \strong{Description} \cr
   \emph{Shape_Leng} \tab Shape length \cr
   \emph{Shape_Area} \tab Shape area \cr
   \emph{ADM1_PT} \tab Administrative area level 1 name in Portuguese \cr
   \emph{ADM1_PCODE} \tab Administrative area level 1 postal code \cr
   \emph{ADM1_REF} \tab Administrative area level 1 reference \cr
   \emph{ADM1ALT1PT} \tab Administrative area level 1 alternative Portuguese name 1 \cr
   \emph{ADM1ALT2PT} \tab Administrative area level 1 alternative Portuguese name 2 \cr
   \emph{ADM0_EN} \tab Administrative area level 0 name in English \cr
   \emph{ADM0_PT} \tab Administrative area level 0 name in Portuguese \cr
   \emph{ADM0_PCODE} \tab Administrative area level 0 postal code \cr
   \emph{date} \tab Date created \cr
   \emph{validOn} \tab Data valid on \cr
   \emph{validTo} \tab Data valid to \cr
}
}
\source{
https://data.humdata.org/dataset/mozambique-administrative-levels-0-3
}
\usage{
provincias
}
\description{
Mozambique \emph{provincias} borders
}
\examples{
provincias

}
\keyword{datasets}
