% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{settlements}
\alias{settlements}
\title{Mozambique settlements locations}
\format{
A \code{SpatialPointsDataFrame} object containing settlements locations of
Mozambique

| \strong{Variables} | \strong{Description} |
| \emph{OBJECTID} | Object identifier |
| \emph{FID_1} | Feature identifier |
| \emph{Sett_Name} | Settlement name |
| \emph{F_CLASS} | Feature class |
| \emph{F_DESIG} | Feature designation |
| \emph{ADM1_Old} | Administrative level 1 name - old |
| \emph{ADM2_Old} | Administrative level 2 name - old |
| \emph{Country} | Country name |
| \emph{ISO3} | Tree letter code for country |
| \emph{ADM1_Name} | Administrative level 1 name - current |
| \emph{ADM1_Code} | Administrative level 1 code - current |
| \emph{ADM1_ID} | Administrative level 1 identifier - current |
| \emph{ADM2_Name} | Administrative level 2 name - current |
| \emph{ADM2_Code} | Administrative level 2 code - current |
| \emph{ADM2_ID} | Administrative level 2 identifier - current |
| \emph{ADM3_Name} | Administrative level 3 name - current |
| \emph{ADM3_Code} | Administrative level 3 code - current |
| \emph{ADM3_ID} | Administrative level 3 identifier - current |
| \emph{ADM3_Type} | Administrative level 3 type - current |
| \emph{ADM3_TypeE} | Administrative level 3 type in English - current |
| \emph{Latitude} | Latitude coordinate |
| \emph{Longitude} | Longitude coordinate |
| \emph{SHAPE_Leng} | Shape length |
}
\source{
https://data.humdata.org/dataset/mozambique-settlement-shapefiles
}
\usage{
settlements
}
\description{
Mozambique settlements locations
}
\examples{
settlements

}
\keyword{datasets}
