% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{villages}
\alias{villages}
\title{Mozambique villages locations}
\format{
A \code{SpatialPointsDataFrame} object containing villages locations of
Mozambique

| \strong{Variables} | \strong{Description} |
| \emph{osm_id} | OpenStreetMap Identifier |
| \emph{name} | Name of location |
| \emph{source} | Source of location data |
| \emph{population} | Population size |
| \emph{is_in} | Is in what country location |
| \emph{place} | Type of location |
}
\source{
https://data.humdata.org/dataset/hotosm_moz_populated_places
}
\usage{
villages
}
\description{
Mozambique villages locations
}
\examples{
villages

}
\keyword{datasets}
